feat: api-handlers-decomposition-003 - Preserve work submission and upsert HTTP behavior
